requirements.txt
django==2.0.4
djangorestframework==3.8.2
channels==2.1.0

